Filter IN with inline if possible
