+++ modules/core/src/main/java/org/apache/ignite/internal/cache/query/index/sorted/inline/InlineIndexKeyType.java
@@ -102,4 +102,15 @@ public interface InlineIndexKeyType {
      * @return Size of key, in bytes. {@code -1} means variable length of key.
      */
     public short keySize();
+
+    /**
+     * Whether inline contains full index key.
+     *
+     * @param pageAddr Page address.
+     * @param off Offset.
+     * @return {@code true} if inline contains full index key. Can be {@code false} for truncated variable lenght types.
+     */
+    public default boolean inlinedFullValue(long pageAddr, int off) {
+        return true;
+    }
 }
